adding aliases to Person
