+++ RecordLib/case.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 from dataclasses import asdict
-from RecordLib.common import Charge, Person, Sentence
+from RecordLib.common import Charge, Sentence
+from RecordLib.person import Person
 from typing import List, Optional
 from datetime import date
 import pytest
