+++ RecordLib/docket/parse_pdf.py
@@ -1,5 +1,6 @@
 from typing import Union, BinaryIO, Tuple, Callable, List, Optional
-from RecordLib.common import Person, Charge, Sentence, SentenceLength
+from RecordLib.common import Charge, Sentence, SentenceLength
+from RecordLib.person import Person
 from RecordLib.case import Case
 from RecordLib.grammars.docket import (
     docket_sections, docket_sections_nonterminals, common_terminals,
