+++ tests/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 from RecordLib.case import Case
-from RecordLib.common import Charge, Person, Sentence, SentenceLength
+from RecordLib.person import Person
+from RecordLib.common import Charge, Sentence, SentenceLength
 from RecordLib.crecord import CRecord
 from RecordLib.attorney import Attorney
 from RecordLib.summary.pdf import parse_pdf as parse_summary_pdf
