+++ tests/test_common.py
@@ -4,36 +4,6 @@ from datetime import date, timedelta
 import pytest
 from RecordLib.serializers import to_serializable
 
-def test_person():
-    per = Person("John", "Smeth", date(2010, 1, 1), date(2020, 1, 1))
-    assert per.first_name == "John"
-    assert per.last_name == "Smeth"
-    assert per.date_of_birth.year == 2010
-    assert per.date_of_death.year == 2020
-
-def test_person_age():
-    per = Person("John", "Smeth", date(2000, 1, 1))
-    assert per.age() > 17
-
-
-def test_person_years_dead(example_person):
-    example_person.date_of_death = None
-    assert example_person.years_dead() == float("-Inf")
-    example_person.date_of_death = date.today()
-    assert example_person.years_dead() == 0
-    example_person.date_of_death = date(2000,1,1)
-    assert example_person.years_dead() > 10
-
-
-def test_person_todict():
-    per = Person("John", "Smeth", date(2010, 1, 1))
-    assert asdict(per) == {
-        "first_name": "John",
-        "last_name": "Smeth",
-        "date_of_birth": date(2010, 1, 1),
-        "date_of_death": None,
-    }
-
 def test_sentence():
     st = Sentence(
         sentence_date=date(2010, 1, 1),
